lwc components, controllers & custom metadata
Added all the components created for Assignment. Created congaObjectSummary, a reusable component which would be fetch summary data based on the record page where it is placed. The Custom Metadata would have a record each for different objects where the component might be used, and the fields__c field on custom metadata would have comma separated field apis which are to be displayed in summary section. The name of the custom metadata record must be the Object API. This makes the component reusable.